+++ messaginginapp/src/androidTest/java/io/customer/messaginginapp/MessagingInAppIntegrationTest.kt
@@ -11,17 +11,17 @@ import io.customer.messaginginapp.gist.presentation.GIST_TAG
 import io.customer.messaginginapp.gist.presentation.GistModalManager
 import io.customer.messaginginapp.gist.presentation.GistSdk
 import io.customer.messaginginapp.provider.GistApi
-import io.customer.messaginginapp.support.GistApiProviderDelegate
-import io.customer.messaginginapp.support.GistEngineMessageDriver.MessageState
-import io.customer.messaginginapp.support.GistEngineWebViewClientInterceptor
-import io.customer.messaginginapp.support.GistServerResponseDispatcher
-import io.customer.messaginginapp.support.awaitWithTimeoutBlocking
-import io.customer.messaginginapp.support.createInAppMessage
-import io.customer.messaginginapp.support.getInAppMessageActivity
-import io.customer.messaginginapp.support.mapToInAppMessage
-import io.customer.messaginginapp.support.pageRuleContains
-import io.customer.messaginginapp.support.pageRuleEquals
-import io.customer.messaginginapp.support.runOnUIThreadBlocking
+import io.customer.messaginginapp.testutils.GistApiProviderDelegate
+import io.customer.messaginginapp.testutils.GistEngineMessageDriver.MessageState
+import io.customer.messaginginapp.testutils.GistEngineWebViewClientInterceptor
+import io.customer.messaginginapp.testutils.GistServerResponseDispatcher
+import io.customer.messaginginapp.testutils.awaitWithTimeoutBlocking
+import io.customer.messaginginapp.testutils.createInAppMessage
+import io.customer.messaginginapp.testutils.getInAppMessageActivity
+import io.customer.messaginginapp.testutils.mapToInAppMessage
+import io.customer.messaginginapp.testutils.pageRuleContains
+import io.customer.messaginginapp.testutils.pageRuleEquals
+import io.customer.messaginginapp.testutils.runOnUIThreadBlocking
 import io.customer.messaginginapp.type.InAppEventListener
 import io.customer.messaginginapp.type.InAppMessage
 import io.customer.sdk.CustomerIO
