+++ messaginginapp/src/androidTest/java/io/customer/messaginginapp/testutils/GistExtensions.kt
@@ -1,4 +1,4 @@
-package io.customer.messaginginapp.support
+package io.customer.messaginginapp.testutils
 
 import android.util.Base64
 import androidx.test.platform.app.InstrumentationRegistry
